services/workers/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
celery[librabbitmq]==5.4.0
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
psycopg2-binary==2.9.9
pyvips==2.2.3
minio==7.2.7
onnxruntime==1.18.0
Pillow==10.4.0
numpy==1.26.4
pika==1.3.2
pydantic-settings==2.3.4
prometheus-fastapi-instrumentator==7.0.0

# Dev / testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0
factory-boy==3.3.0
moto[s3]==5.0.11
testcontainers[postgres]==4.7.2
